+++ src/main/java/ui/gui/DisplayType.java
@@ -1,0 +1,5 @@
+package ui.gui;
+
+public enum DisplayType {
+    HOME, CLI, NONE;
+}
