+++ src/main/java/ui/gui/MainGui.java
@@ -25,21 +25,11 @@ public class MainGui extends Application {
     @Override
     public void start(Stage stage) {
         try {
-            Stage stageTwo = new Stage();
-            FXMLLoader loaderCliDisplay = new FXMLLoader(MainGui.class
-                    .getResource("/view/CommandLineDisplay.fxml"));
-            AnchorPane cliDisplayRoot = loaderCliDisplay.load();
-            CommandLineDisplay cliDisplayController = loaderCliDisplay.<CommandLineDisplay>getController();
-            cliDisplayController.setStyle();
-            Scene sceneCliDisplay = new Scene(cliDisplayRoot);
-            stageTwo.setScene(sceneCliDisplay);
-            stageTwo.setTitle("Command Line Display");
-            stageTwo.show();
 
             FXMLLoader loaderMain = new FXMLLoader(MainGui.class
                     .getResource("/view/MainWindow.fxml"));
             AnchorPane anchorPane = loaderMain.load();
-            loaderMain.<MainWindow>getController().initialize(this.taskPath, this.walletPath, cliDisplayController);
+            loaderMain.<MainWindow>getController().initialize(this.taskPath, this.walletPath);
             Scene scene = new Scene(anchorPane);
             stage.setScene(scene);
             stage.setTitle("Duke$$$");
